feat: Allow admin login via URL param and secret click
Introduces a new way to access the admin login page. Users can now navigate to the login page with a `?admin=true` query parameter in the URL.

Additionally, a "secret" Easter egg is added to the standard login page. Clicking the app logo five times will redirect the user to the admin login page. This provides alternative access routes for administrators.